+++ SampleProjects/TestSomething/test/eeprom.cpp
@@ -1,8 +1,9 @@
 #include <ArduinoUnitTests.h>
 #include <Arduino.h>
-#include <EEPROM.h>
 
-#ifdef EEPROM_SIZE
+// Only run EEPROM tests if there is hardware support!
+#if defined(EEPROM_SIZE) || (defined(E2END) && E2END)
+#include <EEPROM.h>
 
 unittest(length)
 {